Change Dynamo Moscow key from "dynamo" to "dinamo" in 2015-16
In the rest of the repo, Dynamo Moscow has the key "dinamo" except for in the 2015-16 rfpl.yml file. When I build the database and include major-league-soccer, Houston Dynamo is listed as playing in the Russian Premier League in 2015-16 instead of Dynamo Moscow and I think this may be why.